Move add trip request to book actions
